Starting code refactor to fit agreed upon infrastructure
