--- src/main/kotlin/models/NomicProblemDetails.kt
@@ -1,6 +1,0 @@
-package models
-
-data class NomicProblemDetails (
-    val error: Boolean,
-    val msg: String
-)
