+++ src/main/kotlin/nomic/data/dtos/Amendment.kt
@@ -1,4 +1,4 @@
-package models.entities
+package nomic.data.dtos
 
 import org.ktorm.entity.Entity
 import org.ktorm.schema.Table
